entity-detection/v1/requirements.txt
Flask==2.3.3
groq==0.9.0
python-dotenv==1.0.1
Werkzeug==2.3.7